requirements.txt
Flask==2.3.3
google-cloud-firestore==2.13.0
google-cloud-bigquery==3.31.0
gunicorn==21.2.0